Инкремент 11

+++ internal/app/shorturl/db/model.go
@@ -3,5 +3,5 @@ package db
 type ShortURL struct {
 	ID        string `json:"id"`
 	OriginURL string `json:"origin_url"`
-	UserID    uint64 `json:"user_id"`
+	UserID    uint32 `json:"user_id"`
 }
